Improve osbrain global configuration process
Load environment variables properly. Use an importable configuration
variable for global settings.
