+++ docs/source/advanced_proxy_handling.rst
@@ -71,9 +71,11 @@ Or when creating a new :class:`Proxy <osbrain.proxy.Proxy>` to the agent:
 
    agent.method_call()  # This is now by default an unsafe method_call()
 
-It is also possible, although totally unadvisable, to change the default proxy
-behavior globally. Setting the ``OSBRAIN_DEFAULT_SAFE`` environment variable to
-``false`` would result in all proxies making unsafe calls by default.
+.. note:: It is also possible, although totally unadvisable, to change the
+   default proxy behavior globally. Setting the ``OSBRAIN_DEFAULT_SAFE``
+   environment variable to ``false`` would result in all proxies making unsafe
+   calls by default. Another option is to change the ``osbrain.config['SAFE']``
+   configuration variable.
 
 
 .. index:: proxy, oneway, concurrent
