+++ docs/source/transport_protocol.rst
@@ -19,7 +19,7 @@ protocols that can be used in osBrain:
   different machines.
 - ``ipc``: common `IPC <https://en.wikipedia.org/wiki/Inter-process_communication>`_. It is the default and the best suited for communication between agents that
   run on the same machine
-- ``inproc``: for in-process communication (between threads). The fastest 
+- ``inproc``: for in-process communication (between threads). The fastest
   protocol, although it is limited to communication between threads that share
   the same :attr:`Agent.context <osbrain.agent.Agent.context>`
 
@@ -33,12 +33,12 @@ Changing the transport
 ======================
 
 It is possible to change the default global transport protocol by setting
-the ``OSBRAIN_DEFAULT_TRANSPORT`` environment variable. So, for example, to
+the ``osbrain.config['TRANSPORT']`` configuration variable. So, for example, to
 set TCP as the default transport, we could set:
 
 .. code-block:: python
 
-   os.environ['OSBRAIN_DEFAULT_TRANSPORT'] = 'tcp'
+   osbrain.config['TRANSPORT'] = 'tcp'
 
 We can also set the default transport that a particular agent should use by
 default by passing the ``transport`` parameter to
@@ -56,3 +56,6 @@ we can still change the transport protocol when binding, passing the
 
    agent = run_agent('a0')
    agent.bind('PULL', transport='inproc')
+
+.. note:: It is also possible to change the global default transport protocol
+   setting the ``OSBRAIN_DEFAULT_TRANSPORT`` environment variable.
